refactor: move type to a centrol places with different file
