+++ src/types/tag.ts
@@ -1,0 +1,3 @@
+export type Tags = {
+  [key: string]: string;
+};
