+++ src/types/vars.ts
@@ -1,0 +1,11 @@
+export type Vars = {
+  [key: string]: unknown;
+};
+
+export type Stage = {
+  [key: string]: string;
+};
+
+export type Stages = {
+  [key: string]: Stage;
+};
